refactored add focus to plugin
