+++ packages/graph/src/plugins/focus/events.ts
@@ -1,0 +1,17 @@
+import { EventMapToEventBus } from '../../events/index.ts';
+
+export type FocusGraphEventMap = {
+  /**
+   * when the set of focused items changes
+   */
+  onFocusChange: (
+    newItemIds: ReadonlySet<string>,
+    oldItemIds: ReadonlySet<string>,
+  ) => void;
+};
+
+type FocusGraphEventBus = EventMapToEventBus<FocusGraphEventMap>;
+
+export const createFocusGraphEventBus = (): FocusGraphEventBus => ({
+  onFocusChange: new Set(),
+});
